removed extra packages, connected correct main.tsx
