+++ packages/server/package.json
@@ -13,25 +13,21 @@
   },
   "dependencies": {
     "compression": "^1.7.4",
-    "cookie-parser": "^1.4.6",
     "cors": "2.8.5",
     "cross-env": "7.0.3",
     "dotenv": "16.0.3",
     "eslint-config-prettier": "8.8.0",
     "express": "4.18.2",
     "http-proxy-middleware": "^2.0.6",
-    "jsesc": "^3.0.2",
     "pg": "8.10.0",
     "prettier": "2.8.8",
     "sirv": "^2.0.2"
   },
   "devDependencies": {
     "@types/compression": "^1.7.2",
-    "@types/cookie-parser": "^1.4.3",
     "@types/cors": "2.8.13",
     "@types/express": "4.17.17",
     "@types/jest": "28.1.8",
-    "@types/jsesc": "^3.0.1",
     "@types/pg": "8.6.6",
     "@typescript-eslint/eslint-plugin": "5.59.1",
     "@typescript-eslint/parser": "5.59.1",
